+++ libft/libft_strconv/src/num/ft_ulltoa_base.c
@@ -6,7 +6,7 @@
 /*   By: pepe <pepe@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 17:07:09 by jpriou            #+#    #+#             */
-/*   Updated: 2019/08/13 17:05:11 by pepe             ###   ########.fr       */
+/*   Updated: 2019/08/15 23:53:08 by pepe             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,15 +16,15 @@
 
 char		*ft_ulltoa_base(unsigned long long n, const char *base)
 {
-	char	*res;
-	int		index;
-	int		len;
-	size_t	baselen;
+	char			*res;
+	unsigned int	index;
+	unsigned int	len;
+	size_t			baselen;
 
 	if (n == 0)
 		return (ft_strdup("0"));
 	baselen = ft_strlen(base);
-	len = ft_logullb(n, (int)baselen);
+	len = ft_logullb(n, baselen);
 	if ((res = malloc((size_t)len + 1)) == NULL)
 		return (NULL);
 	index = len - 1;
